refactor forms to watch and update the correct fields
`list-form.tsx` component is refactored to use the watch method to render the bookmarks and pass their unique id.

`bookmark-form.tsx` component is refactored to use `setValue` instead of `update` to prevent unmount/remount so that the modal can have it's proper state control. Also sets a random string when `append` is called to avoid duplicate keys in the list form component.