add `session-context.tsx` component to handle refresh on session change
this is a very sketchy workaround to fix an issue where tabs opened in protected routes weren't being refreshed when the user signs out on another tab.

i believe that this is due to the fact that my protected routes are server components, so when a client event happens in another tab, we need a client component wrapped around it to refresh the route when the session changes in the client.

to fix this, i've added a session-context component and wrapped around the pages in `layout.tsx`. this component calls `useSession` and runs useEffect to refresh the page when the session changes in the client.